Moved the `PlaygroundQuickLook` extension out of LogEntry+Reflection.swift and into its own file.
